MAINT: Moved ABC classes and core functions to base.py
